fixing code quality issues
